Add Less and custom Bootstrap
The experiments without Twitter's Bootstrap are removed and a custom
Bootstrap build is added to the project. There is only one CSS file left
with all styles. This file is compiled from Less code including
everything from Bootstrap, that is used in the page. The compiled CSS is
also added to Git, since it is required for Github pages to host the
page.