Make copyright footer more tiny
The copyright text in the footer is now much more tiny. The text is
smaller and in a lighter color. It has some distance to the above
content and the below end of the browser window. The license name has
been shorten to cc-by-sa to keep the footer's text shorter.